refactor all files to packages
