+++ src/main/java/duke/Parser.java
@@ -1,3 +1,13 @@
+package duke;
+
+import duke.command.*;
+import duke.exception.DukeException;
+import duke.exception.DukeIllegalArgumentsException;
+import duke.exception.DukeUnknownCommandException;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Todo;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
