+++ src/main/java/duke/Storage.java
@@ -1,3 +1,11 @@
+package duke;
+
+import duke.exception.DukeStorageException;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.Todo;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
