+++ src/main/java/duke/command/ExitCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.exception.DukeStorageException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
