+++ src/main/java/duke/command/MarkCommand.java
@@ -1,14 +1,21 @@
-public class UnmarkCommand extends Command {
+package duke.command;
+
+import duke.exception.DukeTaskNotFoundException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
+public class MarkCommand extends Command {
     private final int taskIndex;
 
-    public UnmarkCommand(int taskIndex) {
+    public MarkCommand(int taskIndex) {
         this.taskIndex = taskIndex;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeTaskNotFoundException {
         try {
-            ui.showUndoneMessage(tasks.getTask(taskIndex).markAsUndone());
+            ui.showDoneMessage(tasks.getTask(taskIndex).markAsDone());
         } catch (IndexOutOfBoundsException e) {
             throw new DukeTaskNotFoundException();
         }
