+++ src/main/java/duke/exception/DukeIllegalArgumentsException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class DukeIllegalArgumentsException extends DukeException {
     public DukeIllegalArgumentsException(String message) {
         super(message);
