+++ src/main/java/duke/exception/DukeStorageException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class DukeStorageException extends DukeException {
     public DukeStorageException(String message) {
         super(message);
