+++ src/main/java/duke/exception/DukeTaskNotFoundException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class DukeTaskNotFoundException extends DukeException {
     public DukeTaskNotFoundException() {
         super("I can't find that task!\n");
