+++ src/main/java/duke/exception/DukeUnknownCommandException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class DukeUnknownCommandException extends DukeException {
     public DukeUnknownCommandException() {
         super("I'm sorry, but I don't know what that means :-(\n");
